+++ src/starplot/__init__.py
@@ -1,3 +1,6 @@
 """Star charts and maps"""
 
 __version__ = "0.1.2"
+
+from .zenith import ZenithPlot  # noqa: F401
+from .map import MapPlot  # noqa: F401
